# contract/src/main/java/io/pravega/schemaregistry/contract/generated/rest/model/Application.java
/*
 * Pravega Schema Registry APIs
 * REST APIs for Pravega Schema Registry.
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.pravega.schemaregistry.contract.generated.rest.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.pravega.schemaregistry.contract.generated.rest.model.SchemaInfoList;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import javax.validation.constraints.*;

/**
 * Application
 */

public class Application   {
  @JsonProperty("name")
  private String name = null;

  @JsonProperty("writingTo")
  private Map<String, SchemaInfoList> writingTo = null;

  @JsonProperty("readingFrom")
  private Map<String, SchemaInfoList> readingFrom = null;

  @JsonProperty("properties")
  private Map<String, String> properties = null;

  public Application name(String name) {
    this.name = name;
    return this;
  }

  /**
   * Get name
   * @return name
   **/
  @JsonProperty("name")
  @ApiModelProperty(value = "")
  public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }

  public Application writingTo(Map<String, SchemaInfoList> writingTo) {
    this.writingTo = writingTo;
    return this;
  }

  public Application putWritingToItem(String key, SchemaInfoList writingToItem) {
    if (this.writingTo == null) {
      this.writingTo = new HashMap<String, SchemaInfoList>();
    }
    this.writingTo.put(key, writingToItem);
    return this;
  }

  /**
   * Get writingTo
   * @return writingTo
   **/
  @JsonProperty("writingTo")
  @ApiModelProperty(value = "")
  public Map<String, SchemaInfoList> getWritingTo() {
    return writingTo;
  }

  public void setWritingTo(Map<String, SchemaInfoList> writingTo) {
    this.writingTo = writingTo;
  }

  public Application readingFrom(Map<String, SchemaInfoList> readingFrom) {
    this.readingFrom = readingFrom;
    return this;
  }

  public Application putReadingFromItem(String key, SchemaInfoList readingFromItem) {
    if (this.readingFrom == null) {
      this.readingFrom = new HashMap<String, SchemaInfoList>();
    }
    this.readingFrom.put(key, readingFromItem);
    return this;
  }

  /**
   * Get readingFrom
   * @return readingFrom
   **/
  @JsonProperty("readingFrom")
  @ApiModelProperty(value = "")
  public Map<String, SchemaInfoList> getReadingFrom() {
    return readingFrom;
  }

  public void setReadingFrom(Map<String, SchemaInfoList> readingFrom) {
    this.readingFrom = readingFrom;
  }

  public Application properties(Map<String, String> properties) {
    this.properties = properties;
    return this;
  }

  public Application putPropertiesItem(String key, String propertiesItem) {
    if (this.properties == null) {
      this.properties = new HashMap<String, String>();
    }
    this.properties.put(key, propertiesItem);
    return this;
  }

  /**
   * Get properties
   * @return properties
   **/
  @JsonProperty("properties")
  @ApiModelProperty(value = "")
  public Map<String, String> getProperties() {
    return properties;
  }

  public void setProperties(Map<String, String> properties) {
    this.properties = properties;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Application application = (Application) o;
    return Objects.equals(this.name, application.name) &&
        Objects.equals(this.writingTo, application.writingTo) &&
        Objects.equals(this.readingFrom, application.readingFrom) &&
        Objects.equals(this.properties, application.properties);
  }

  @Override
  public int hashCode() {
    return Objects.hash(name, writingTo, readingFrom, properties);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Application {\n");
    
    sb.append("    name: ").append(toIndentedString(name)).append("\n");
    sb.append("    writingTo: ").append(toIndentedString(writingTo)).append("\n");
    sb.append("    readingFrom: ").append(toIndentedString(readingFrom)).append("\n");
    sb.append("    properties: ").append(toIndentedString(properties)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}