# contract/src/main/java/io/pravega/schemaregistry/contract/generated/rest/model/ApplicationsInGroup.java
/*
 * Pravega Schema Registry APIs
 * REST APIs for Pravega Schema Registry.
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.pravega.schemaregistry.contract.generated.rest.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.pravega.schemaregistry.contract.generated.rest.model.VersionInfoList;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import javax.validation.constraints.*;

/**
 * ApplicationsInGroup
 */

public class ApplicationsInGroup   {
  @JsonProperty("map")
  private Map<String, VersionInfoList> map = null;

  public ApplicationsInGroup map(Map<String, VersionInfoList> map) {
    this.map = map;
    return this;
  }

  public ApplicationsInGroup putMapItem(String key, VersionInfoList mapItem) {
    if (this.map == null) {
      this.map = new HashMap<String, VersionInfoList>();
    }
    this.map.put(key, mapItem);
    return this;
  }

  /**
   * Get map
   * @return map
   **/
  @JsonProperty("map")
  @ApiModelProperty(value = "")
  public Map<String, VersionInfoList> getMap() {
    return map;
  }

  public void setMap(Map<String, VersionInfoList> map) {
    this.map = map;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    ApplicationsInGroup applicationsInGroup = (ApplicationsInGroup) o;
    return Objects.equals(this.map, applicationsInGroup.map);
  }

  @Override
  public int hashCode() {
    return Objects.hash(map);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class ApplicationsInGroup {\n");
    
    sb.append("    map: ").append(toIndentedString(map)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}