# contract/src/main/java/io/pravega/schemaregistry/contract/generated/rest/model/SchemaInfoList.java
/*
 * Pravega Schema Registry APIs
 * REST APIs for Pravega Schema Registry.
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.pravega.schemaregistry.contract.generated.rest.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.pravega.schemaregistry.contract.generated.rest.model.SchemaInfo;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.ArrayList;
import java.util.List;
import javax.validation.constraints.*;

/**
 * SchemaInfoList
 */

public class SchemaInfoList   {
  @JsonProperty("schemas")
  private List<SchemaInfo> schemas = null;

  public SchemaInfoList schemas(List<SchemaInfo> schemas) {
    this.schemas = schemas;
    return this;
  }

  public SchemaInfoList addSchemasItem(SchemaInfo schemasItem) {
    if (this.schemas == null) {
      this.schemas = new ArrayList<SchemaInfo>();
    }
    this.schemas.add(schemasItem);
    return this;
  }

  /**
   * Get schemas
   * @return schemas
   **/
  @JsonProperty("schemas")
  @ApiModelProperty(value = "")
  public List<SchemaInfo> getSchemas() {
    return schemas;
  }

  public void setSchemas(List<SchemaInfo> schemas) {
    this.schemas = schemas;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    SchemaInfoList schemaInfoList = (SchemaInfoList) o;
    return Objects.equals(this.schemas, schemaInfoList.schemas);
  }

  @Override
  public int hashCode() {
    return Objects.hash(schemas);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class SchemaInfoList {\n");
    
    sb.append("    schemas: ").append(toIndentedString(schemas)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}