# contract/src/main/swagger/SchemaRegistry.yaml
...
swagger: "2.0"
info:
  description: "REST APIs for Pravega Schema Registry."
  version: "0.0.1"
  title: Pravega Schema Registry APIs
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0"
basePath: "/v1"
...
schemes:
  - http
paths:
...
  /applications:
    post:
      tags:
      - "Group"
      operationId: createApplication
      description: Create a new Application
      consumes:
        - application/json
      parameters:
        - in: body
          name: CreateApplicationRequest
          description: The Group configuration
          required: true
          schema:
            type: object
            properties:
              applicationName:
                type: string
              properties:
                type: object
                additionalProperties:
                  type: string
                  minLength: 0
                  maxLength: 40
      responses:
        201:
          description: Successfully added group
        409:
          description: Group with given name already exists
        500:
          description: Internal server error while creating a Group
  /applications/{appId}:
    parameters:
      - in: path
        name: appId
        description: appId
        required: true
        type: string
    get:
      tags:
      - "Group"
      operationId: getApplication
      description: Fetch the application
      produces:
        - application/json
      responses:
        200:
          description: Found Group properties
          schema:
            $ref: "#/definitions/Application"
        404:
          description: Group with given name not found
        500:
          description: Internal server error while fetching Group details
  /applications/{appId}/reader:
    parameters:
      - in: path
        name: appId
        description: appId
        required: true
        type: string
    post:
      tags:
      - "Application"
      operationId: addReader
      description: Add new reader to application
      consumes:
        - application/json
      parameters:
        - in: body
          name: AddReaderRequest
          description: schema version
          required: true
          schema:
            type: object
            properties:
              groupId:
                type: string
              version:
                  $ref: "#/definitions/VersionInfo"
      responses:
        201:
          description: Successfully added reader 
        409:
          description: Schema incompatible
        500:
          description: Internal server error while adding reader to group
  /applications/{appId}/writer:
    parameters:
      - in: path
        name: appId
        description: appId
        required: true
        type: string
    post:
      tags:
      - "Application"
      operationId: addWriter
      description: Add new writer to application
      consumes:
        - application/json
      parameters:
        - in: body
          name: AddWriterRequest
          description: schema version
          required: true
          schema:
            type: object
            properties:
              groupId:
                type: string
              version:
                  $ref: "#/definitions/VersionInfo"
      responses:
        201:
          description: Successfully added writer 
        409:
          description: Schema incompatible
          schema:
            $ref: "#/definitions/IncompatibleSchema"

        500:
          description: Internal server error while adding writer to group
  /applications/groups/{groupId}/readers:
    parameters:
      - in: path
        name: groupId
        description: groupId
        required: true
        type: string
    get:
      tags:
      - "Application"
      operationId: listReaders
      description: List all writer applications in group
      produces:
        - application/json
      responses:
        200:
          description: List of all writers in group
          schema:
            $ref: "#/definitions/ApplicationsInGroup"
        500:
          description: Internal server error while fetching the list of writers in Groups
  /applications/groups/{groupId}/writers:
    parameters:
      - in: path
        name: groupId
        description: groupId
        required: true
        type: string
    get:
      tags:
      - "Application"
      operationId: listWriters
      description: List all writer applications in group
      produces:
        - application/json
      responses:
        200:
          description: List of all writers in group
          schema:
            $ref: "#/definitions/ApplicationsInGroup"
        500:
          description: Internal server error while fetching the list of writers in Groups
definitions:
...
  SchemaType:
    type: object
    properties:
      schemaType:
        type: string
        enum: 
          - None
          - Avro
          - Protobuf
          - Json
          - Any
          - Custom
      customTypeName:
        type: string  
  SchemaInfo:
    type: object
    properties:
      schemaName:
        type: string
      schemaType:
          $ref: "#/definitions/SchemaType"
      schemaData:
        type: string
        format: binary
      properties:
        type: object
        additionalProperties:
          type: string
          minLength: 0
          maxLength: 40
  VersionInfo:
    type: object
    properties:
      schemaName:
        type: string
      version:
        type: integer
        format: int32
...
  ApplicationsInGroup:
    type: object
    properties:
      map:
        type: object
        additionalProperties:
          $ref: '#/definitions/VersionInfoList'
  IncompatibleSchema:
    type: object
    properties:
      errorMessage:
        type: string        
  VersionInfoList:
    type: object
    properties:
      versions:
        type: array
        items:
          $ref: "#/definitions/VersionInfo"
  SchemaInfoList:
    type: object
    properties:
      schemas:
        type: array
        items:
          $ref: "#/definitions/SchemaInfo"
  Application:
    type: object
    properties:
      name:
        type: string
      writingTo:
        type: object
        additionalProperties:
          $ref: '#/definitions/SchemaInfoList'
      readingFrom:
        type: object
        additionalProperties:
          $ref: '#/definitions/SchemaInfoList'
      properties:
        type: object
        additionalProperties:
          type: string